Fixed Alignment Oversight
Changed the assembler definition to include a pad instruction
for the cases when a 32bit instruction is made to straddle across
two words of memory. As a result also had to handle the pad
instruction in the CPU by ignoring it (a maskable NOP).
Renamed the arguments of the IO callback function prototype to
not conflict with the write system call.
Made the CPU state-printing function more prettier by adding
the convention names of the registers.