Preliminary Telnet IO
Added io/telnet.h and io/telnet.c that implement a serial port over the telnet protocol.
The telnet IO device exposed to the CPU is hard-coded to appear on port 1 for now.
The emulator now waits for a telnet connection to be made by a client for it to continue.
Further improvements are planned, especially its move to a separate thread.