Interrupts But Working (Barely)
- Changed the HS format to HG by adding a register address.
- Added the $jp special purpose register for holding the interrupt return address.
- Added three new instructions for managing the new $jp:
Two move instructions from loading $jp from a datareg and storing $jp to a datareg.
One "resume" instruction for jumping to $jp and de-escalating privileges.

- Added a new CPU component: the Interrupts & Privileged Mode (IPM) component,
which is responsible for multiplexing the interrupts and holding the privilege state.
It is currently unfinished and only provides basic functionality.
- IPM provides a new IO port called isrloc, which is a loopback from a CPU parameter
called start_addr. This IO port controls the entry point for privileged code.